Use separate focus state for optimization
